Added python script tool to convert normal Draw functions to enable partial drawing while spawning

+++ Engine/StarBackground.cpp
@@ -12,8 +12,8 @@ StarBackground::StarBackground(std::mt19937 & rng, Vec2 & vel_in)
 
 	while (i < nStars)
 	{
-		xStarPoints[i] = xDist(rng);
-		yStarPoints[i] = yDist(rng);
+		xStarPoints[i] = (float)xDist(rng);
+		yStarPoints[i] = (float)yDist(rng);
 		i++;
 	}
 }
@@ -22,14 +22,13 @@ void StarBackground::Draw(Graphics& gfx) const
 {	
 	for (int i = 0; i < nStars; i++)
 	{
-		SpriteCodex::DrawStars(gfx, xStarPoints[i], yStarPoints[i]);
+		SpriteCodex::DrawStars(gfx, (int)xStarPoints[i], (int)yStarPoints[i]);
 	}
 }
 
 void StarBackground::UpdateStars(std::mt19937& rng, const float dt)
 {
 	std::uniform_int_distribution<int> xDist(2, Graphics::ScreenWidth - 3);
-	//std::uniform_int_distribution<int> yDist(2, Graphics::ScreenHeight - 3);
 
 	for (int i = 0; i < nStars; i++)
 	{
@@ -38,8 +37,8 @@ void StarBackground::UpdateStars(std::mt19937& rng, const float dt)
 
 		if (xStarPoints[i] < 2 || xStarPoints[i] > Graphics::ScreenWidth - 3 || yStarPoints[i] < 2 || yStarPoints[i] > Graphics::ScreenHeight - 3)
 		{
-			xStarPoints[i] = xDist(rng);
-			yStarPoints[i] = Graphics::ScreenHeight - 3;
+			xStarPoints[i] = (float)xDist(rng);
+			yStarPoints[i] = (float)(Graphics::ScreenHeight - 3);
 		}
 	}
 }
